u-boot-toradex: add bbappend for U-Boot 2019.07
Add TorizonCore specific bbappend for u-boot-toradex 2019.07. This
appends patches which enable distro boot on raw NAND devices as
well as makes sure that the internal eMMC device is the first boot
device (since TorizonCore is meant to be a production ready distro
it is porbably not a good idea to boot from an external boot
medium first by default).

Related-to: #52582
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>