meta: introduce ostree-kernel-initramfs
A build error was observed as follows:
| DEBUG: Executing shell function do_image_ostree
| cat: imx6dl-colibri-eval-v3.dtb: No such file or directory
| cp: cannot create regular file 'usr/lib/modules/5.4.18-0.0.0-devel+git.0e2cde6/vmlinuz': No such file or directory

it's not always reproducible, but I reproduced it once when I was
building master and zeus branches in a same work directory.

It's caused by the kernel version mismatched in rootfs directory and in
do_image_ostree which is reading KERNEL_VERSION.

It could be fixed if we package kernel image, devicetrees, initramfs
and install them to rootfs, but with this solution, we will have a
annoying QA warning as follows:
| WARNING: ostree-kernel-initramfs-0.0.1-r0 do_package: Manifest ...initramfs-ostree-torizon-image.packagedata
| not found in colibri_imx6 armv7ahf-neon-imx armv7at2hf-neon-imx armv7at2hf-neon armv7ahf-neon armv7at2hf-vfp
| armv7ahf-vfp armv6thf-vfp armv6hf-vfp armv5tehf-vfp armv5ehf-vfp armv5thf-vfp armv5hf-vfp allarch
| x86_64_x86_64-nativesdk (variant '')?

this is due to the initramfs recipe has do_packagedata task but there
is no package manifest generated, to avoid the warning, we could
let it inherit nopackages.

Signed-off-by: Ming Liu <ming.liu@toradex.com>