torizon-machine-custom: remove export DTC_FLAGS
Remove the export of DTC_FLAGS from the torizon-machine-custom class.
The goal of the export was to add the "-@" argument which compiles
symbols into the devicetree which is necessary for devicetree overlays.
However, it did not work properly. The mainline kernel does this by
default. For the downstream kernel DTC_FLAGS are now set directly in the
bbappend file of the kernel.

Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>