u-boot-distro-boot: reset upgrade_available in the rollback case
Make sure upgrade_available is set back to 0 when rolling back. This
avoids unnecessary writes to the U-Boot environment when staying in
rollbacked mode. It also matches the state diagram in the aktualizr
documentation.

We have to implement the check at the top to avoid saving lots of
temporary environment variables.

Also protect the expressions passed to test with quotes. This is how
meta-lmp implemented the tests and allows to get rid of one test.

Related-to: TOR-955
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
(cherry picked from commit 4443ad8357636141c6ed1c3a6018cb09eed7231d)