linux-toradex-mainline: append to BSP recipe
Instead of maintain our own, append to the BSP recipe. This makes sure
that TorizonCore releases align with the BSP also in terms of the
mainline based kernel.

Note that this recipe can also be built with PREEMPT_RT support when the
preempt-rt override is set (which is the case in the torizon-rt
distribution). The BSP changed to use the same recipe for RT builds for
all machines with the change ac5028651745 ("machines: rt mainline kernel:
use the common linux-toradex-mainline"). Since the recipe TorizonCore
provided was not setup to build a RT kernel, the TorizonCore RT images
have been shipped with a regular kernel. Since with this change we
inherit the RT capable linux-toradex-mainline, this also fixes the RT
image builds to ship with a PREEMPT_RT kernel again.

Related-to: TOR-1152
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
(cherry picked from commit f1ac2d9828860c7f2a6be4692830d640e124b92c)