u-boot-distro-boot: move reset_upgrade_available logic to the boot script
The reset_upgrade_available logic saves the environment, and being
inside uEnv.txt it means it will save all environment defined in the
uEnv.txt file. That is really not necessary, and has been causing
problems with modules where the U-Boot environment size is <= 8K. So
let's move this logic to the boot script.

Related-to: TOR-1610

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>