setup-environment-internal: fix issue when manifest.xml is not a symlink
The repo tool has been changed a while ago (see [1]) to create the
manifest file (.repo/manifest.xml) as a plain xml file instead of a
symlink.

This impacts the logic that checks if we are building from default.xml,
and decide if use-head-next should be added to MACHINEOVERRIDES.

Change this logic to support the manifest.xml file as both a symlink and
a regular xml file.

[1] https://android.googlesource.com/tools/repo/+/a269b1cb9dc21dfd598bfea3766206b606ad4589

Related-to: TOR-1883

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>