mwifiex-delay: fix issue when hostname is changed
Since the script uses the hostname to identify the SoM, it fails when
it's changed by the user.

So let's fix that by changing the script to use the MACHINE environment
variable. This variable is created when building the image, it will
always have the module name and should not be changed by the user.

Related-to: TOR-2199

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>