qemuarm64: fix graphics issue
For graphics to work on qemuarm64, QEMU must be built with SDL support.
Add a few lines in the local.conf template file so the user can enable
it.

This is not enabled by default because a native build of SDL depends on
having x11 in DISTRO_FEATURES, and we unconditionally remove it in our
distro configuration file. This is solved by ASSUME_PROVIDED +=
"libsdl2-native", forcing the usage of libsdl2 installed in the host
machine, but then this could break our CI builds.

So for now we can just leave QEMU builds with graphics disabled by
default, and manually enable it when required.

Related-to: TOR-2455

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>


(cherry picked from commit feacfb289f5f9e034d0df3ce28ae4efc178ca61a)