u-boot-ota: fix issue when extracting default environment
LTO was enbaled for u-boot, and this changed the name of the section the
default environment was being stored in the binary.
With this, a new application was added to u-boot to extract that
environment, so we just need to change our recipe to use it.

This change was ported to our downstream u-boot, so we need to use this
new application. But for upstream, we're still using and older version
of u-boot, which doesn't have this change yet, so check for the
availability of 'printiinitialenv' application, and if not preset we use
the old method to extract it.

For reference, this is the change made into u-boot to use the new
application to extract the environment:
https://source.denx.de/u-boot/u-boot/-/commit/486aef08de091ca35386f32fe961a201c3cfa9d4

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>