u-boot: conditionally disable container generation by U-Boot
Disable binman so that the bootloader container is not produced by the
build of U-Boot. This is done only on the i.MX8M machines since binman
is enabled only for them by default (at the time of writing) and it is
conditional on UBOOT_PROVIDES_BOOT_CONTAINER to give users a choice.
With secure boot we foresee setting that variable to "0" to use the
container generation from imx-boot so that all the signing logic from
meta-toradex-security can work (before we adapt it to binman).

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>