uEnv.txt: store valid value in fuse_val_N upon fuse read errors
Originally the code responsible for reading fuse values and storing them
into the fuse_val_<N> variable would store any random value in memory
into that variable in case of a fuse read error. Here we fix this by
detecting the error and storing a value of 0x0 in those cases.

Also, change the logic to store the values into fuse_val_<N> already
with a prefix of 0x to denote the fact that they are in hex format.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>