uEnv.txt: delete fuse_val_{num} variable upon read errors
In case a fuse value cannot be read, delete the corresponding variable
rather than setting it to 0x0. This avoids passing potentially wrong
information to the OS.

An example of a situation where the information would be wrong before:
if a closed device was running an old bootloader with hardening enabled
then the command for reading the fuses would be denied and with the old
logic all of the fuse_val_{num} variables would be set to 0x0 even
though the fuses have other values (notice that since the device is
closed for sure the fuses have been programmed).

The code on the OS side can use the absence of the fuse_val_{num}
variables as an indication that the bootloader is missing support for
the fusing feature.

Related-to: TOR-3657

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>