uEnv.txt: prevent fuse reading when bootloader support is missing
Add logic to detect if the bootloader has support for the fuse reading
feature. The decision is based on information provided by the command
"tdx_secboot_get" (or by its absence). If the command does not exist we
know for sure the bootloader is old or does not have the required
patches for the fuse reading code to work. The required patches would
include:

- The addition of the "tdx_secboot_get" command itself to provide
  build and runtime information about the running bootloader.
- The addition of the command "fuse readv" for reading the fuses
  securely into a variable.
- The addition of the new -q switch to the various fuse commands
  to make them quiet (avoid bloating the boot logs).

Related-to: TOR-3657

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>