am62xx: make bbappend changes only apply to TI's AM62 builds
We have some 'meta-ti-bsp' custom .bbappends on dynamic-layers, but the
issue here is that Toradex's HW also uses this layer so we end up
applying those configurations as well.

In the end, it resulted in 'fw_printenv' not working properly, since the
configuration file for TI is different from Toradex, which caused
greenboot to not work properly and the OS rollback process to fail
altogether.

Now we add a new override for TI's AM62 devices in our
'common-am62xx.inc' file, namely 'common-ti', since there's not unique
override on TI's BSP, and 'ti-soc' that was used was common to both
BSP's. So now, on TI's specific .bbappends, we use the 'common-ti'
override, to make sure we're using Common Torizon distro and TI machine.

Related-to: TOR-3777

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>
(cherry picked from commit 14856900e0c1aaf8ad9ab6b5bb60d0206ce6980d)
Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>