u-boot-rollback: remove CONFIG_BOOTCOMMAND override in Torizon
There's a legacy u-boot fragment that is overriding the 'bootcommand'.
So far it wasn't causing any issues, since we were using a valid
bootcommand. But with Aquila AM69, it is using 'bootflow' instead of
distroboot, so this override was causing a boot issue.

Since this is something we get from Toradex's BSP, we can safely remove
the bootcommand, but we leave the fragment there, if we need this for
a device supported by Common Torizon.

Related-to: TOR-3800

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>
(cherry picked from commit 5b9869b9fb34bde565d8cbb3c35046ecd1534083)