meta-synaptics:synsdk-gpu-pvr: fix PREBUILT_* variable paths
The recipe 'synsdk-gpu-pvr', specifies 2 variables to point to paths
where we can find some prebuilts, PREBUILT_LIBS and PREBUILT_BINS.

The issue here is that this variable is constructed using HOST_SYS
variable, which takes the form: {HOST_ARCH}-{HOST_VENDOR}-{HOST_OS},
which on Torizon, this would become 'aarch64-tzn-linux'.

The issue here comes from the fact that we're using prebuilts, and we're
downloading them from the Synaptics repository
'sysroot-linux-baseline'[1], and there we only have 1 type of prebuilt,
with a hardcoded path, which was built with a different HOST_SYS:
aarch64-poky-linux.

So for Torizon, we need to manually change the PREBUILT_* variable to no
longer use HOST_SYS, and pass manually the correct path.

Also create the dynamic-layer for meta-synaptics.

[1] https://github.com/synaptics-astra/sysroot-linux-baseline/tree/scarthgap_6.12_v2.0.0_alpha/data/gfx_prebuilt/imagination/wayland/sl1680

Related-to: TOR-3945

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>
(cherry picked from commit 6bebf61aa920049101d152fe2fbfb848fef12e34)
Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>