u-boot-distro-boot: Add new method for applying custom kernel args
A new method for applying custom kernel args is required, because
the old method does not work with FIT images which are important
for root-of-trust purposes.

It should be noted the old method is still present and supported,
however the new method will take priority if both are in use.

Related-to: TOR-4066

Signed-off-by: Jeremias Cordoba <jeremias.cordoba@toradex.com>
(cherry picked from commit deb5e339fb47930156a22398c2f6a6b29b3654c4)