u-boot-distro-boot: add ability to prepend custom kernel arguments
Replace the previous custom kernel arguments setting method, which only
supported appending, with a new method that supports both prepending and
appending. The variables torizon_bootargs_l and torizon_bootargs_r are
prepended and appended, respectively, to bootargs, which U-Boot passes
to the Linux kernel at runtime.

The idea behind this is to give TorizonCore Builder or other tooling the
flexibility to decide whether to prepend or append based on features of
the image. In particular, prepending would be used to handle Secure Boot
images having the kernel command-line protection feature enabled.

Also fix a bug where the custom bootargs string would not roll back to
the value from the previous deployment when that previous value was
empty.

Related-to: TCB-763
Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>
(adapted from commit e7742b4bf87600dc586c85273d9698cc911fcda8)