meta-beagle:linux: enable Toshiba TC358768 MIPI DSI bridge
After bumping meta-ti, we started facing an issue on BeagleY AI builds,
where initramfs was failing to build due to it no finding the module for
TC358768.
Seems that after the bump, we this config was no longer set as module,
but as 'not set', causing this failure.
So now we add this config to our fragment, as a module, to fix the build
issue.

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>
(cherry picked from commit b8c0211382810c79c1b406c0913a7b9edb6a88ee)
Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>